digitalsys_api/requirements.txt
amqp==5.1.1
asgiref==3.4.1
awscli==1.22.14
billiard==4.1.0
boto3==1.20.14
botocore==1.23.14
celery==5.3.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.3
Django==3.2.9
django-cors-headers==3.10.0
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
docutils==0.15.2
drf-yasg==1.21.6
gunicorn==20.1.0
inflection==0.5.1
jmespath==0.10.0
kombu==5.3.1
mysqlclient==2.1.0
packaging==23.1
Pillow==8.4.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pyasn1==0.4.8
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.6